solve invalid call exception

+++ bomber.py
@@ -16,10 +16,14 @@ async def call(message: aiogram.types.Message):
 async def spam(message: aiogram.types.Message, state: FSMContext):
     await state.finish()
     text = message.text
-    for i in range(10):
-        msg = await message.answer(text)
-        time.sleep(5)
-        await msg.delete()
+    if text.startswith('@'):
+        for i in range(10):
+            msg = await message.answer(text)
+            time.sleep(5)
+            await msg.delete()
+    else:
+        await message.answer('Призыв должен начинатся с @ [@ivanzolo2004]')
+        await Form.spam_it_state.set()
 
 
 class Form(StatesGroup):
